machine: Add machine type for skylake with tpm2-tis & txt.
This new machine type is for my dell 7050 test system. Adding tpm2-tis &
txt to the MACHINE_FEATURES of the existing intel-skylake-64 machine
should only require including an existing conf and appending the strings
to the variable (see intel-corei7-64_tpm-txt from this layer).
Unfortunately the tune-skylake.inc from meta-intel disables qemu-usermode
which makes it incompatible with the read-only-rootfs image feature.
This in turn breaks core-image-tpm which uses read-only-rootfs.

I've been unable to come up with a better way (one that doesn't depend
on weird things like the order of BBPATH) to override the use of
MACHINE_FEATURES_remove in the skylake tune from meta-intel and so this
commit copies the tune config locally as tune-skylake-rofs.inc. The
intel-skylake-64_tpm2-tis-txt.conf machine is similarly based on the
intel-skylake-64 machine from meta-intel. The only local change beyond
adding the tpm2-tis and txt features is the inclusion of the skylake
tune compatible with read-only-rootfs.

Signed-off-by: Philip Tricca <flihp@twobit.org>